Re-write rewards-distribution-token using new sdk-core (#833)
