--- examples/rewards-distribution-token/.env.template
@@ -1,2 +1,0 @@
-GOERLI_MNEMONIC=rigid cradle south ...
-GOERLI_PROVIDER_URL=https://goerli.infura.io/v3/<API_KEY>
