+++ examples/rewards-distribution-token/.eslintignore
@@ -1,0 +1,4 @@
+node_modules
+artifacts
+cache
+coverage
