+++ examples/rewards-distribution-token/.npmignore
@@ -1,0 +1,3 @@
+hardhat.config.ts
+scripts
+test
