+++ examples/rewards-distribution-token/.prettierignore
@@ -1,0 +1,5 @@
+node_modules
+artifacts
+cache
+coverage*
+gasReporterOutput.json
