+++ examples/rewards-distribution-token/.prettierrc
@@ -1,0 +1,1 @@
+{}
