+++ examples/rewards-distribution-token/README.md
@@ -1,31 +1,46 @@
-Rewards Distribution Token
-==========================
+# Advanced Sample Hardhat Project
 
-A ERC20 token that tokenizes units in the [Instant Distribution Agreements](https://docs.superfluid.finance/superfluid/protocol-tutorials/perform-an-instant-distribution/).
+This project demonstrates an advanced Hardhat use case, integrating other tools commonly used alongside Hardhat in the ecosystem.
 
-All holders of the rights tokens can receive cash token rewards in one distribution transaction in two ways:
+The project comes with a sample contract, a test for that contract, a sample script that deploys that contract, and an example of a task implementation, which simply lists the available accounts. It also comes with a variety of other tools, preconfigured to work with the project code.
 
-- If the subscription to the token is approved by its holder, then the distribution is available the moment distribution is done,
-- if not, then the pending distribution could be claimed by its holder any time in a separated transaction.
+Try running some of the following tasks:
 
-# Usage
-
-## Run tests
-
-```bash
-yarn install
-yarn build
-yarn test
+```shell
+npx hardhat accounts
+npx hardhat compile
+npx hardhat clean
+npx hardhat test
+npx hardhat node
+npx hardhat help
+REPORT_GAS=true npx hardhat test
+npx hardhat coverage
+npx hardhat run scripts/deploy.ts
+TS_NODE_FILES=true npx ts-node scripts/deploy.ts
+npx eslint '**/*.{js,ts}'
+npx eslint '**/*.{js,ts}' --fix
+npx prettier '**/*.{json,sol,md}' --check
+npx prettier '**/*.{json,sol,md}' --write
+npx solhint 'contracts/**/*.sol'
+npx solhint 'contracts/**/*.sol' --fix
 ```
 
-## Deployment
+# Etherscan verification
 
-Setup a `.env` file following the example of `.env.template`.
+To try out Etherscan verification, you first need to deploy a contract to an Ethereum network that's supported by Etherscan, such as Ropsten.
 
-```bash
-RELEASE_VERSION=v1 npx truffle --network goerli exec scripts/deploy.js
+In this project, copy the .env.example file to a file named .env, and then edit it to fill in the details. Enter your Etherscan API key, your Ropsten node URL (eg from Alchemy), and the private key of the account which will send the deployment transaction. With a valid .env file in place, first deploy your contract:
+
+```shell
+hardhat run --network ropsten scripts/deploy.ts
 ```
 
-## UI
+Then, copy the deployment address and paste it in to replace `DEPLOYED_CONTRACT_ADDRESS` in this command:
 
-A minimal one-pager plain HTML/jS ui is available also under `ui/` folder.
+```shell
+npx hardhat verify --network ropsten DEPLOYED_CONTRACT_ADDRESS "Hello, Hardhat!"
+```
+
+# Performance optimizations
+
+For faster runs of your tests and scripts, consider skipping ts-node's type checking by setting the environment variable `TS_NODE_TRANSPILE_ONLY` to `1` in hardhat's environment. For more details see [the documentation](https://hardhat.org/guides/typescript.html#performance-optimizations).
