--- examples/rewards-distribution-token/test/.eslintrc.js
@@ -1,19 +1,0 @@
-module.exports = {
-    globals: {
-        web3: true,
-        artifacts: true,
-        describe: true,
-        contract: true,
-        context: true,
-        before: true,
-        after: true,
-        beforeEach: true,
-        afterEach: true,
-        it: true,
-        assert: true
-    },
-    rules: {
-        "no-console": "off",
-        indent: "off"
-    }
-};
