+++ examples/rewards-distribution-token/ui/index.html
@@ -29,7 +29,7 @@
   <script>if (typeof module === "undefined") module = {};</script>
   <script src="https://cdn.jsdelivr.net/npm/@superfluid-finance/js-sdk@latest/src/abi.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/@superfluid-finance/js-sdk@latest/src/getConfig.js"></script>
-  <script src="DividendRightsToken.abi.js"></script>
+  <script src="../test/abis/fDAIABI.js"></script>
   <script>
     let web3, daix, user, host, ida, drt;
